telegram-bot/requirements.txt
aiogram==3.3.0
kafka-python==2.0.2
minio==7.1.0
python-dotenv==0.19.0
Pillow==10.2.0